refactor: Grayscale effect to core.
